'driver' interface for backing up files
Once the backup file has been generated it'll be handled by a 'driver'.
Drivers are responsible for backup transfers. This design pattern allows
m2bk to transfer backups via a handful of possibilities, not only s3 but
also anything that can be abstracted as a driver (SFTP, Dropbox, Google
Drive, etc.)